Update Lottery.sol
Final smart-contract.

Deploy with time desired for length of draw in seconds, parameter - drawLengthInSeconds
e.g. 604800 for 1 week

If no one enters in specified length, the draw will be extended on the next entry

When the draw expires, the next address that enters will still be entered into the draw. The smart-contract will then commence random number generation along with winnings distribution, then restart draw with a new endTime relative to drawLengthInSeconds.

*Further fail-safes and precautionary code added.